Move user form to different component
